Finalizes the unstyled pop-up
