+++ components/index.js
@@ -1,5 +1,5 @@
 import InfoBar from "./InfoBar";
-import PopUp from "./PopUp";
+import PopupContents from "./PopupContents";
 import NavBar from "./NavBar";
 import SearchBar from "./SearchBar";
 import MapContainer from "./MapContainer";
@@ -13,7 +13,7 @@ import BurgerMenu from "./BurgerMenu";
 
 export {
     InfoBar,
-    PopUp,
+    PopupContents,
     NavBar,
     SearchBar,
     MapContainer,
