+++ package.json
@@ -26,6 +26,7 @@
 		"react-dom": "^16.6.3",
 		"react-places-autocomplete": "^7.2.0",
 		"react-responsive": "^5.0.0",
+		"reactjs-popup": "^1.3.1",
 		"save": "^2.3.2",
 		"webpack": "^4.20.2"
 	},
